project(data-lovers): Remove visual identity paragraph
Although the files that were shared somehow could help the students to have an idea of what to expect of this project. Using `.sketch` or `.psd` files could restrict their ability to access  those resources (not everyone has the propietary software to open them). And while there was one that had a `.pdf` extension, we could create a bias towards using that data-set instead of the others without a visual identity.

Let's take a step back and rethink the benefits of providing these files and the format in which we can share them.